realtime-service/requirements.txt
fastapi
uvicorn
python-socketio
aiohttp
python-dotenv
